renamed stuff to javazone

+++ src/test/java/nl/craftsmen/crypto/KeystoreTest.java
@@ -23,18 +23,19 @@ import java.math.BigInteger;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.security.*;
+import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.Locale;
+import java.util.*;
 
 public class KeystoreTest {
 
 
     @Test
     public void createKeyStore() throws GeneralSecurityException, IOException, OperatorCreationException {
-        Files.delete(Path.of("myStore.pkcs12"));
+        if (Files.exists(Path.of("myStore.pkcs12"))) {
+            Files.delete(Path.of("myStore.pkcs12"));
+        }
         KeyStore keyStore = KeyStore.getInstance("PKCS12");
         keyStore.load(null, null);
         //create keypair
@@ -45,7 +46,16 @@ public class KeystoreTest {
         Key pvt = kp.getPrivate();
         X509Certificate x509Certificate = generateRootCert(kp);
         keyStore.setCertificateEntry("myRootCertificate", x509Certificate);
+        Certificate[] chain = {x509Certificate};
+        keyStore.setKeyEntry("myPrivateKey", pvt, "password".toCharArray(), chain);
         keyStore.store(new FileOutputStream("myStore.pkcs12"), "password".toCharArray());
+        Enumeration<String> enums = keyStore.aliases();
+        System.out.println("aliases:");
+        enums.asIterator().forEachRemaining(System.out::println);
+        Key key = keyStore.getKey("myPrivateKey", "password".toCharArray());
+        System.out.println(key.toString());
+        Certificate certificate = keyStore.getCertificate("myRootCertificate");
+        System.out.println(certificate.toString());
     }
 
     public X509Certificate generateRootCert(KeyPair pair) throws IOException, OperatorCreationException, CertificateException {
